Ajusta arquivos
