--- src/pages/Tab4.tsx
@@ -1,15 +1,0 @@
-import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar } from '@ionic/react';
-import ExploreContainer from '../components/ExploreContainer';
-import './Tab4.css';
-
-const Tab4: React.FC = () => {
-  return (
-    <IonPage>
-      <IonContent fullscreen>
-        <ExploreContainer name="Tab 4 page" />
-      </IonContent>
-    </IonPage>
-  );
-};
-
-export default Tab4;
