+++ src/routes/AppRoutes.jsx
@@ -1,12 +1,14 @@
 import { Redirect, Route } from "react-router-dom";
 
-import Login from "../pages/Login";
-import  SignupPage from "../pages/Signup"; 
-import AuthenticatedPage from "../pages/AuthenticatedPage";
+import Login from "../pages/Login/index";
+import  SignupPage from "../pages/Signup/index"; 
+import AuthenticatedPage from "../pages/Authenticated/index";
 import { IonApp, IonRouterOutlet } from "@ionic/react";
 import { IonReactRouter } from "@ionic/react-router";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { auth } from "../services/firebaseConfig";
+import Tab1 from "../pages/Tab1";
+import Tab2 from "../pages/Tab2";
 
 export function AppRoutes() {
     const [ user, loading, error ] = useAuthState(auth);
@@ -22,6 +24,12 @@ export function AppRoutes() {
                         {user ? <Redirect to="/" /> : <SignupPage /> }
                     </Route>
                     <Route exact path="/">
+                        {user ? <AuthenticatedPage /> : <Login /> }
+                    </Route>
+                    <Route exact path="/tab1">
+                        {user ? <AuthenticatedPage /> : <Login /> }
+                    </Route>
+                    <Route exact path="/tab2">
                         {user ? <AuthenticatedPage /> : <Login /> }
                     </Route>
                 </IonRouterOutlet>
